guest stubs: add overwritable indirection
If an app overwrites symbols in NDK libs our trampolines bound to these
symbols are invalidated. This change attempts to bind trampolines to
hidden symbols instead. These symbols are not discoverable by the app and
thus better protected from invalidation.

As a matter of optimization we still bind exported symbols to
trampolines as before. The hidden symbols only take into action
if regular ones are invalidated.

Flag: EXEMPT NDK
Test: the affected app works
Test: tree-hugger
Bug: 378772009
Change-Id: I07cef1c7dcb4c91c00b72dfa4624ec8fac3791c8