Add __BIONIC_DISABLE_MALLOC_USABLE_SIZE_FORTIFY_WARNINGS to allow _FORTIFY_SOURCE=3 with malloc_usable_size
We plan to make _FORTIFY_SOURCE=3 global on device and the header files
use malloc_usable_size. In this case, it is trampoline function for
native_bridge_malloc_usable_size. We found the cc file that includes
these headers and defined
__BIONIC_DISABLE_MALLOC_USABLE_SIZE_FORTIFY_WARNINGS to disable compiler
error.

I added the TODO that you can pull the define out into a separte file.
The other definitions might not get any use out of _FORTIFY_SOURCE so
it might not be worth it.

Flag: EXEMPT bugfix
Bug: 291762537
Test: m with ag/32745870
Change-Id: I5f2f77efec85cfb239cd9e48b69b8b212f814953